tipc: Eliminate warnings associated with print buffer size values
commit cff98cd23b01893ecdb1c7d4bd1869cd6b2f6600 from
git://tipc.cslab.ericsson.net/pub/git/people/allan/tipc.git

Reworks TIPC's print buffer code to record the buffer size using "size_t",
rather than "int". This prevents warnings about potential loss of high
order significant bits when the results of C library string manipulation
routines (which use "size_t") are used by print buffer routines.

Integrated-by: Howard Xu <hao.xu@windriver.com>

+++ net/tipc/tipc_bcast.c
@@ -734,7 +734,7 @@ void tipc_bcbearer_push(void)
 
 #ifdef CONFIG_TIPC_CONFIG_SERVICE
 
-int tipc_bclink_stats(char *buf, const u32 buf_size)
+size_t tipc_bclink_stats(char *buf, const size_t buf_size)
 {
 	struct print_buf pb;
